Time: 29 ms (43.16%), Space: 92.1 MB (43.94%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.java
@@ -2,7 +2,7 @@ class Solution {
     private boolean isPossible(int[] candies, long k, int divisor) {
         long childrens = 0;
         for(int c: candies) {
-            childrens += (long)c/divisor;
+            childrens += c/divisor;
             if(childrens>=k) return true;
         }
         return false;
